clean up duplicate test file
